Resolve conflicts arising from rebasing updated swift-api
